[#79] Moved all the functions in `stalker/models/__init__.py` to the `stalker.models.utils` module.
